# api.yaml
openapi: 3.0.3
info:
  title: API Системы учета и выдачи инструментов
  description: |
    Спецификация REST API для синхронизации, авторизации и управления реестром оборудования с проверкой Ed25519 подписей оффлайн.
    Подробная логика — в документе «Описание проекта подробно, используемое в разработке».

    ## Время
    Все метки времени в API и в протоколе QR — Unix time в секундах (UTC, int64); от часового пояса не зависят.
    Клиенты работают по времени сервера: `serverTime` из `/api/auth/server-key` и `/api/auth/worker-cert`
    задаёт смещение локальных часов.

    ## Авторизация
    JWT передаётся в заголовке `Authorization: Bearer <accessToken>`. Срок действия — 12 часов.
    Содержит id пользователя (claim `sub`), роль (claim `role`) и id его компании (claim `company_id`);
    подписан HMAC-SHA256 (секрет `JWT_SIGNING_KEY`).
    Отзыва нет: после увольнения выданный JWT действует до истечения срока.

    ## Идентификаторы
    id пользователей, компаний и инструментов — UUID в каноническом виде (строчные hex, `8-4-4-4-12`). Их нельзя
    перебрать и по ним нельзя оценить число пользователей или инструментов. Защитой данных они не являются:
    от доступа к чужому объекту защищает проверка компании (см. ниже), а не секретность id.
    Значение, не являющееся UUID, в пути или запросе — 400.
    id записей журнала (`logId`, `RentalLogEntry.id`) — обычные числа: на вход они нигде не принимаются,
    записи всегда выбираются по компании.

    ## Компании
...

...

...

...

...

...
servers:
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      operationId: Register
      summary: Регистрация нового пользователя
      description: |
        Регистрируются все роли. Компания задаётся по роли:
        - **Owner** передаёт `companyName` — создаётся новая компания, а Owner становится её первым пользователем
          и подтверждается сразу (is_approved = true): подтверждать некому. `companyId` не передаётся.
        - **Worker** и **Issuer** передают `companyId` существующей компании (id сообщает её Owner, он виден в
          `GET /api/companies/my`); аккаунт создаётся с is_approved = false и ждёт подтверждения Owner этой компании.
          `companyName` не передаётся.

        Во всех случаях is_active = true. Телефон нормализуется к `+7XXXXXXXXXX` и уникален среди всех пользователей,
        включая уволенных: `89991234567` и `+7 (999) 123-45-67` — один и тот же номер. Без авторизации.

        Если регистрация не удалась (409), компания не создаётся.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      security: []
      responses:
        '201':
          description: |
            Аккаунт зарегистрирован (без тела). Owner может входить сразу; Worker и Issuer — после подтверждения Owner.
        '400':
          description: |
            Некорректный запрос (ValidationProblemDetails). Помимо обычных ошибок полей:
            - поля компании не соответствуют роли: у Owner нет `companyName` или передан `companyId`;
              у Worker и Issuer нет `companyId` или передан `companyName`;
            - `companyId` — компании с таким id нет.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.1
                title: One or more validation errors occurred.
                status: 400
                errors:
                  companyId: ['Компания с таким id не найдена']
                traceId: 00-2f1c...-01
        '409':
          description: Пользователь с таким телефоном уже существует
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.10
                title: Conflict
                status: 409
                detail: Пользователь с таким телефоном уже существует
                traceId: 00-2f1c...-01
  /api/auth/login:
    post:
      tags: [Auth]
      operationId: Login
      summary: Вход в систему
      description: |
        Возвращает JWT (срок действия 12 часов; содержит id пользователя, роль и id его компании).
        Вход разрешён только при is_approved = true и is_active = true.
        Телефон перед поиском нормализуется так же, как при регистрации. Без авторизации.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      security: []
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Неверный телефон или пароль
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.2
                title: Unauthorized
                status: 401
                detail: Неверный телефон или пароль
                traceId: 00-2f1c...-01
        '403':
          description: Аккаунт не подтверждён (is_approved = false) или уволен (is_active = false)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.4
                title: Forbidden
                status: 403
                detail: Аккаунт ещё не подтверждён владельцем
                traceId: 00-2f1c...-01
...
  /api/auth/worker-cert:
    get:
      tags: [Auth]
      operationId: GetWorkerCert
      summary: Выдача сертификата и временного ключа сотруднику
      description: |
        Сервер генерирует временную пару ключей Ed25519, формирует токен ServerSignedToken
        (workerId, fio, issuedAt, expiresAt = issuedAt + 12 часов, workerPublicKey), подписывает его своим
        приватным ключом и отдаёт сотруднику вместе с временным приватным ключом. Приватный ключ сервером
        не сохраняется. Выдаётся только при is_approved = true и is_active = true (проверяется по БД в момент запроса).

        **Требуется роль:** Worker.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Сертификат (token + serverSignature), временный приватный ключ и время сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerCertificate'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: |
            Роль не Worker — без тела (проверка JWT, как у Forbidden);
            аккаунт не подтверждён или уволен (is_approved = false или is_active = false, проверка по БД в момент
            запроса) — тело ProblemDetails.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.4
                title: Forbidden
                status: 403
                detail: Сотрудник уволен, сертификат не выдаётся
                traceId: 00-2f1c...-01

  # ==========================================
  # КОМПАНИИ
  # ==========================================
  /api/companies/my:
    get:
      tags: [Companies]
      operationId: GetMyCompany
      summary: Своя компания
      description: |
        Id и название компании вызывающего. Owner сообщает этот id сотрудникам и завхозам, они указывают его
        при регистрации. Чужую компанию получить нельзя, как и список всех компаний. Отдельного создания
        компании нет: она создаётся при регистрации Owner.

        **Требуется авторизация:** любая роль.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Компания вызывающего
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==========================================
  # УПРАВЛЕНИЕ ПОЛЬЗОВАТЕЛЯМИ
  # ==========================================
...
  /api/sync:
    post:
      tags: [Sync]
      operationId: Sync
      summary: Синхронизация оффлайн-выдач и возвратов
      description: |
        Завхоз отправляет массив записей своего локального журнала: строка QR сотрудника как есть
        и время сканирования. Сервер сам строго разбирает каждую строку по формату AT1 (см. OfflineRentalPayload),
        проверяет обе подписи, время, инструмент и повтор, и сохраняет **каждую** запись в журнал независимо
        от результата этих проверок — результат пишется в validationFlag (по первой непройденной проверке,
        порядок — см. схему ValidationFlag).

        Строка `qr` не по формату AT1 — ошибка конверта, как и неверное тело запроса: честный клиент завхоза
        такую строку не примет при оффлайн-проверке. Запрос целиком отклоняется (400), ничего не сохраняется.

        issuer_id записи берётся из JWT вызывающего.

        **Идемпотентность:** запись с тем же issuer_id, scannedAt и той же строкой qr, уже сохранённая
        ранее (повторная отправка после обрыва связи), повторно не создаётся — в ответе возвращается существующая.

        Запись принадлежит компании завхоза (company_id из JWT). Сотрудник из токена и инструмент из операции
        ищутся только в этой компании: сотрудник другой компании неотличим от несуществующего (UNKNOWN_WORKER),
        инструмент другой компании — тоже (UNKNOWN_TOOL). Привязки worker_id и tool_id у записи заполняются
        только проверенным и найденным в компании завхоза, иначе null; чужие id в журнал не попадают.

        При флаге VALID сервер обновляет `tools.condition` значением `toolCondition`, если время операции
        (timestamp) не раньше последнего изменения состояния инструмента (>=).

        Пустой массив допустим — ответ `[]`.

        **Требуется роль:** Issuer.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/SyncRecord'
      responses:
        '200':
          description: Все записи сохранены; результаты в порядке записей запроса
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SyncResult'
              example:
                - logId: 41
                  validationFlag: VALID
                - logId: 42
                  validationFlag: TIME_DRIFT
        '400':
          description: |
            Тело запроса не является массивом объектов SyncRecord, либо хотя бы одна строка `qr` не соответствует
            формату AT1. Ничего не сохраняется. Для записей с неверным форматом ключ ошибки — номер записи
            в массиве (с 0): `"[3].qr"`. По этим номерам клиент завхоза откладывает такие записи (локально,
            для разбора) и отправляет остальные повторно, чтобы одна запись не блокировала синхронизацию всей очереди.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.1
                title: One or more validation errors occurred.
                status: 400
                errors:
                  '[3].qr': ['Строка не соответствует формату AT1: workerSignature должна быть 64 байта']
                  '[7].qr': ['Строка не соответствует формату AT1: ожидается 5 частей через точку']
                traceId: 00-2f1c...-01
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/logs:
    get:
      tags: [Sync]
      operationId: GetLogs
      summary: Просмотр журнала движения инструментов
      description: |
        Возвращает записи rental_logs своей компании (записи, синхронизированные её завхозами), новые сверху:
        по убыванию времени операции (qrTimestamp), затем по убыванию id.
        Записи со всеми флагами, не только VALID.

        **Требуется роль:** Issuer, Owner.
      security:
        - BearerAuth: []
      parameters:
        - name: action
          in: query
          required: false
          description: Фильтр по типу операции
          schema:
            $ref: '#/components/schemas/RentalAction'
        - name: toolId
          in: query
          required: false
          description: Фильтр по инструменту (записи с toolId = null в выборку не попадают); не UUID — 400
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: История операций
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RentalLogEntry'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT из POST /api/auth/login, срок действия 12 часов
...
  responses:
    ValidationError:
      description: |
        Некорректный запрос: тело не разбирается как JSON нужной формы, нарушены обязательность или ограничения полей,
        неверный параметр пути или запроса. Ключи `errors` — пути к полям в именах JSON.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblemDetails'
          example:
            type: https://tools.ietf.org/html/rfc9110#section-15.5.1
            title: One or more validation errors occurred.
            status: 400
            errors:
              fullName: ['The field FullName must be a string or array type with a maximum length of ''100''.']
              role: ['The Role field is required.']
            traceId: 00-2f1c...-01
    Unauthorized:
      description: Не передан, просрочен или недействителен JWT. Без тела; заголовок `WWW-Authenticate` со схемой Bearer
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Bearer error="invalid_token", error_description="The token expired at '...'"
    Forbidden:
      description: |
        Роль пользователя не допускается к этому эндпоинту. Без тела.
        Увольнение уже выданный JWT не отзывает (см. раздел «Авторизация»).
    NotFound:
      description: Объект с таким id не найден (в том числе принадлежащий другой компании)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://tools.ietf.org/html/rfc9110#section-15.5.5
            title: Not Found
            status: 404
            detail: Инструмент с id 17c4e5a2-93b8-4f10-8d6a-0b2e7c9d4f35 не найден
            traceId: 00-2f1c...-01
  schemas:
...
    ProblemDetails:
      type: object
      description: |
        Описание ошибки в формате RFC 9457 (ранее RFC 7807), `application/problem+json`.
        Формирует ASP.NET Core; допускает дополнительные поля.
      required: [type, title, status]
      properties:
        type:
          type: string
          description: URI типа ошибки (для стандартных кодов — ссылка на раздел RFC 9110)
          example: https://tools.ietf.org/html/rfc9110#section-15.5.5
        title:
          type: string
          description: Краткое название типа ошибки (стандартное, на английском)
          example: Not Found
        status:
          type: integer
          format: int32
          description: HTTP-код ответа
          example: 404
        detail:
          type: string
          nullable: true
          description: Пояснение к конкретному случаю на русском, для показа пользователю
          example: Инструмент с id 17c4e5a2-93b8-4f10-8d6a-0b2e7c9d4f35 не найден
        instance:
          type: string
          nullable: true
          description: URI конкретного случая ошибки (обычно не заполняется)
        traceId:
          type: string
          description: Идентификатор запроса для поиска в логах сервера
          example: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-00
      additionalProperties: true
    ValidationProblemDetails:
      description: ProblemDetails с перечнем ошибок по полям (ответ 400)
      allOf:
        - $ref: '#/components/schemas/ProblemDetails'
        - type: object
          required: [errors]
          properties:
            errors:
              type: object
              description: |
                Ключ — путь к полю в именах JSON (`fullName`, `[3].qr`, `id`) или JSON-путь при ошибке разбора
                (`$.role`); значение — список сообщений об ошибках.
              additionalProperties:
                type: array
                items:
                  type: string
              example:
                '[3].qr': ['Строка не соответствует формату AT1: ожидается 5 частей через точку']

    # ---------- Перечисления ----------
    UserRole:
      type: string
      description: Worker — сотрудник, Issuer — завхоз, Owner — владелец (регистрируются все три)
      enum: [Worker, Issuer, Owner]
    ToolCondition:
      type: string
      description: Состояние инструмента, по возрастанию тяжести (Good < Damaged < Broken)
      enum: [Good, Damaged, Broken]
    RentalAction:
      type: string
      description: TAKE — сотрудник берёт инструмент, GIVE — возвращает
      enum: [TAKE, GIVE]
...
    RegisterRequest:
      type: object
      description: |
        Поля компании зависят от роли: Owner передаёт `companyName` (и не передаёт `companyId`),
        Worker и Issuer — `companyId` (и не передают `companyName`). Иначе — 400.
      required: [fullName, phone, password, role]
      properties:
        fullName:
          type: string
          maxLength: 100
          description: ФИО. Попадает в QR сотрудника (fio в token), поэтому ограничено 100 символами
          example: "Иванов Иван Иванович"
        phone:
          type: string
          pattern: '^[\s()-]*(\+7|7|8)(?:[\s()-]*[0-9]){10}[\s()-]*$'
          description: |
            Телефон — логин. Любой допустимый формат (см. раздел «Телефон»); сохраняется как `+7XXXXXXXXXX`.
            Уникален среди всех пользователей, включая уволенных
          example: "+7 (999) 123-45-67"
        password:
          type: string
          format: password
          example: "securePassword123"
        role:
          $ref: '#/components/schemas/UserRole'
        companyId:
          type: string
          format: uuid
          description: |
            Только Worker и Issuer, обязателен для них. id существующей компании: его сообщает Owner
            (виден ему в GET /api/companies/my)
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        companyName:
          type: string
          maxLength: 100
          description: |
            Только Owner, обязателен для него. Название новой компании (не уникально).
            Owner становится первым пользователем компании и подтверждается сразу
          example: "ООО Ромашка"
    LoginRequest:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
          pattern: '^[\s()-]*(\+7|7|8)(?:[\s()-]*[0-9]){10}[\s()-]*$'
          description: Любой допустимый формат (см. раздел «Телефон»); перед поиском нормализуется
          example: "89991234567"
        password:
          type: string
          format: password
          example: "securePassword123"
    LoginResponse:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string
          description: "JWT, срок действия 12 часов; передаётся в заголовке `Authorization: Bearer <accessToken>`"
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1Iiwicm9sZSI6Iklzc3VlciJ9.c2lnbmF0dXJl
...
    WorkerCertificate:
      type: object
      description: |
        Временный сертификат и ключ, которые работник получает, пока есть интернет.
        Сертификат — это пара token + serverSignature. Содержимое token — ServerSignedToken
        (клиент декодирует base64url → JSON). Для подписи операций token и serverSignature
        используются как строки, в точности как получены.
      required: [token, serverSignature, temporaryPrivateKey, serverTime]
      properties:
        token:
          type: string
          description: base64url(UTF-8 JSON ServerSignedToken)
        serverSignature:
          type: string
          description: base64url(Ed25519(ключ сервера, ASCII "AT1-TOKEN." + token)), 64 байта (86 символов)
        temporaryPrivateKey:
          type: string
          description: |
            Временный приватный ключ Ed25519 (32-байтный seed), base64url без паддинга (43 символа).
            В WebCrypto импортируется как JWK { kty: "OKP", crv: "Ed25519", d: temporaryPrivateKey, x: workerPublicKey из token }.
        serverTime:
          type: integer
          format: int64
          description: Текущее время сервера (Unix time) — для вычисления смещения часов клиента
          example: 1790000000
    SyncRecord:
      type: object
      description: Запись локального журнала завхоза
      required: [qr, scannedAt]
      properties:
        qr:
          $ref: '#/components/schemas/OfflineRentalPayload'
        scannedAt:
          type: integer
          format: int64
          description: Время сканирования по часам завхоза с поправкой на смещение относительно сервера (Unix time). Не подписывается — за него отвечает завхоз (issuer_id).
          example: 1790003605
    SyncResult:
      type: object
      required: [logId, validationFlag]
      properties:
        logId:
          type: integer
          format: int32
          description: id записи в журнале (для повторно отправленной записи — id уже существующей)
          example: 41
        validationFlag:
          $ref: '#/components/schemas/ValidationFlag'

    # ---------- Основные модели ----------
    Company:
      type: object
      description: Компания. Кроме названия других полей нет
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
          description: Неугадываемый id; по нему сотрудники и завхозы присоединяются к компании при регистрации
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        name:
          type: string
          maxLength: 100
          description: Название (не уникально)
          example: "ООО Ромашка"
...
    RentalLogEntry:
      type: object
      description: |
        Запись в неизменяемом журнале операций; принадлежит компании завхоза (issuerId). Привязка заполняется
        только проверенным и найденным в компании завхоза, иначе null:
        - workerId — id из токена, если подпись сервера верна (поддельный токен не приписывает операцию сотруднику)
          и сотрудник есть в компании завхоза; null при INVALID_SERVER_SIG и UNKNOWN_WORKER;
        - toolId — toolId из операции, если подпись работника верна и инструмент есть в реестре компании завхоза;
          null при INVALID_SERVER_SIG, INVALID_WORKER_SIG, UNKNOWN_TOOL;
        - workerName/toolName берутся из users/tools, если workerId/toolId заполнены, иначе — из заявленных
          непроверенных значений внутри строки QR (fio, toolName).
      required: [id, toolId, toolName, workerId, workerName, issuerId, action, toolCondition, qrTimestamp, scannedAt, validationFlag, createdAt]
      properties:
        id:
          type: integer
          format: int32
          example: 41
        toolId:
          type: string
          format: uuid
          nullable: true
          example: 17c4e5a2-93b8-4f10-8d6a-0b2e7c9d4f35
        toolName:
          type: string
          example: "Дрель Bosch GSB 16 RE"
        workerId:
          type: string
          format: uuid
          nullable: true
          example: 5b1d0c8a-2f4e-4a63-9d17-8e0c3a6f4b21
        workerName:
          type: string
          example: "Иванов Иван Иванович"
        issuerId:
          type: string
          format: uuid
          description: id завхоза, синхронизировавшего запись (из его JWT); его компания — компания записи
          example: 2a6f9d34-81c5-4b07-a3e2-6d5c7f1e9b08
        action:
          $ref: '#/components/schemas/RentalAction'
        toolCondition:
          $ref: '#/components/schemas/ToolCondition'
        qrTimestamp:
          type: integer
          format: int64
          description: timestamp из операции (время операции)
          example: 1790003600
        scannedAt:
          type: integer
          format: int64
          description: Время сканирования завхозом
          example: 1790003605
        validationFlag:
          $ref: '#/components/schemas/ValidationFlag'
        createdAt:
          type: integer
          format: int64
          description: Время сохранения записи на сервере (синхронизации), Unix time
          example: 1790010000